refactor: consolidate the logger interface family via type aliases (#2486)
