+++ docs/shared-utils.md
@@ -167,6 +167,8 @@ This catalog only covers **cross-cutting** helpers — formatters, error helpers
 | ------------------------------- | ------------------------------------------------------- | ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
 | `server/system/logger/index.ts` | `log.{error, warn, info, debug}(namespace, msg, data?)` | All server-side logging. **Never** `console.*` directly. The first arg is a short namespace string (`"accounting"`, `"wiki"`, …) matching the existing convention so log filtering works. |
 | `packages/core/src/host/hostSlot.ts` | `createHostSlot<T>(name)` → `{ set, get, peek, reset }`, `createForwardingLogger(getLogger)` | The host-injection seam inside `@mulmoclaude/core`. A slot holds one host-supplied dependency wired once at boot: `get()` throws (with `name`) if unread, `peek()` stays quiet for non-critical reads, `set()` rejects a re-bind to a different value. `createForwardingLogger` builds the `(prefix, msg, data?)` proxy that drops calls before the host is wired. Used by the collection/feeds/google `host.ts` bindings — reuse it for any new engine binding instead of re-copying the `let current; configure; require; throw` skeleton (#2401). Core-internal (not a package export). |
+| `@mulmoclaude/common`           | `StructuredLogger` (type)                               | **The one declaration** (#2486) of the 4-method `(prefix, message, data?)` host-logger shape. Host `Logger`, core's `hostSlot` + `CollectionLogger` / `FeedsLogger` / `GoogleLogger`, and accounting's `AccountingLogger` are all `export type X = StructuredLogger` aliases — a domain that needs a public logger name aliases this, never re-declares the shape. (`@mulmobridge/chat-service`'s `Logger` stays local: that package has no `@mulmoclaude/common` dep.) |
+| `@mulmoclaude/common`           | `MinimalLogger` (type)                                  | **The one declaration** (#2486) of the 3-method pre-namespaced `(message, data?)` logger shape. `SchedulerLogger` and mulmoscript's `MulmoScriptServerLog` alias it; `NotifierLogger` / `CollectionWatcherLogger` are `Pick<MinimalLogger, …>` subsets. |
 
 ## i18n
